redact admin files to apps
