+++ supply/admin.py
@@ -1,7 +1,6 @@
-from django.contrib import admin  # type: ignore
+from django.contrib import admin
 
 from .models import SupplySwitch
-from users.models import BaseUser
 
 
 @admin.register(SupplySwitch)
@@ -20,30 +19,3 @@ class SwitchAdmin(admin.ModelAdmin):
         "location",
     )
     search_fields = ("name",)
-
-
-@admin.register(BaseUser)
-class UsersAdmin(admin.ModelAdmin):
-    list_display = (
-        "id",
-        "last_login",
-        "is_superuser",
-        "username",
-        "first_name",
-        "last_name",
-        "is_staff",
-        "is_active",
-        "date_joined",
-        "groups",
-        "user_permissions",
-        "preview",
-        "email",
-        "phone_number",
-    )
-    list_filter = (
-        "id",
-        "last_login",
-        "is_superuser",
-        "groups",
-    )
-    search_fields = ("username",)
